refactoring: 인기 검색어 TOP 10 조회 변경
- 전체 기간의 count가 아니라 로그 기반의 count based로 인기 검색어를 조회 하도록 변경
- window size의 경우 설정값으로 조정 가능하며 default는 2000으로 했음

+++ src/main/java/com/github/asyu/homework/domain/trending/dto/response/TrendingDto.java
@@ -2,6 +2,6 @@ package com.github.asyu.homework.domain.trending.dto.response;
 
 public record TrendingDto(
     String query,
-    Integer queryCount
+    Long queryCount
 ) {
 }
